backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
pypdf==3.17.4
python-docx==1.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.4.0
python-dotenv==1.0.0
aiofiles==23.2.1
sentence-transformers==2.3.1
docx2txt==0.8
